01/pyproject.toml
[tool.poetry]
name = "ouchi-se-01"
version = "0.1.0"
description = ""
authors = ["Ayaka Ikezawa <hello@ikeay.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
schedule = "^1.2.2"
pushbullet-py = "^0.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
